feat: added getScorecardUrl as utility
