+++ src/components/ProjectDetails.tsx
@@ -1,5 +1,6 @@
 import { useParams } from "react-router-dom";
 import { useQuery } from "@tanstack/react-query";
+import { getScorecardUrl } from "../utils/getScorecardUrl";
 
 import "./ProjectDetails.css";
 interface ScoreElement {
@@ -13,24 +14,9 @@ interface ScoreElement {
   };
 }
 
-interface paramsTypes {
-  platform: string | undefined;
-  org: string | undefined;
-  repo: string | undefined;
-  commitHash: string | undefined;
-}
-
 function ProjectDetails() {
   const params = useParams();
   const { platform, org, repo, commitHash } = params;
-
-  const getScorecardUrl = (params: paramsTypes): string => {
-    let baseUrl = `https://api.securityscorecards.dev/projects/${params.platform}/${params.org}/${params.repo}`;
-    if (params.commitHash) {
-      baseUrl += `/?commit=${params.commitHash}`;
-    }
-    return baseUrl;
-  };
 
   const { isLoading, error, data } = useQuery({
     queryKey: ["projectData"],
